test[next]: move tests around and add feature markers (#2643)
