+++ tests/next_tests/integration_tests/feature_tests/dace_tests/test_orchestration.py
@@ -33,6 +33,7 @@ from next_tests.integration_tests.multi_feature_tests.ffront_tests.test_laplacia
 dace = pytest.importorskip("dace")
 
 
+@pytest.mark.uses_cartesian_shift
 def test_sdfgConvertible_laplap(cartesian_case):  # noqa: F811
     if not cartesian_case.backend or "dace" not in cartesian_case.backend.name:
         pytest.skip("DaCe-related test: Test SDFGConvertible interface for GT4Py programs")
