+++ tests/next_tests/integration_tests/feature_tests/ffront_tests/test_arg_call_interface.py
@@ -7,15 +7,14 @@
 # SPDX-License-Identifier: BSD-3-Clause
 
 import itertools
-import re
 import typing
 
 import numpy as np
 import pytest
 
+from gt4py import next as gtx
 from gt4py.next import errors
 from gt4py.next.ffront.decorator import field_operator, program, scan_operator
-from gt4py.next.ffront.fbuiltins import broadcast, int32
 
 from next_tests.integration_tests import cases
 from next_tests.integration_tests.cases import IDim, IField, IJKFloatField, KDim, cartesian_case
@@ -206,108 +205,6 @@ def test_call_scan_operator_from_program(cartesian_case):
     )
 
 
-@pytest.mark.uses_scan
-def test_scan_wrong_return_type(cartesian_case):
-    with pytest.raises(
-        errors.DSLError,
-        match=(
-            r"Argument 'state' to scan operator 'testee_scan' must have same type as its return"
-        ),
-    ):
-
-        @scan_operator(axis=KDim, forward=True, init=0)
-        def testee_scan(state: int32) -> float:
-            return 1.0
-
-        @program
-        def testee(qc: cases.IKFloatField, param_1: int32, param_2: float, scalar: float):
-            testee_scan(qc, param_1, param_2, scalar, out=(qc, param_1, param_2))
-
-
-@pytest.mark.uses_scan
-def test_scan_wrong_init_type(cartesian_case):
-    with pytest.raises(
-        errors.DSLError,
-        match=(
-            r"Argument 'init' to scan operator 'testee_scan' must have same type as 'state' argument"
-        ),
-    ):
-
-        @scan_operator(axis=KDim, forward=True, init=0)
-        def testee_scan(state: float) -> float:
-            return 1.0
-
-        @program
-        def testee(qc: cases.IKFloatField, param_1: int32, param_2: float, scalar: float):
-            testee_scan(qc, param_1, param_2, scalar, out=(qc, param_1, param_2))
-
-
-@pytest.mark.uses_scan
-def test_scan_without_carry(cartesian_case):
-    with pytest.raises(
-        errors.DSLError,
-        match=r"Scan operator 'testee_scan' must have at least one argument",
-    ):
-
-        @scan_operator(axis=KDim, forward=True, init=0)
-        def testee_scan() -> float:
-            return 1.0
-
-
-@pytest.fixture
-def bound_args_testee():
-    @field_operator
-    def fieldop_bound_args() -> cases.IField:
-        return broadcast(0, (IDim,))
-
-    @program
-    def program_bound_args(arg1: bool, arg2: bool, out: cases.IField):
-        # for the test itself we don't care what happens here, but empty programs are not supported
-        fieldop_bound_args(out=out)
-
-    return program_bound_args
-
-
-def test_bind_invalid_arg(cartesian_case, bound_args_testee):
-    with pytest.raises(
-        TypeError, match="Keyword argument 'inexistent_arg' is not a valid program parameter."
-    ):
-        bound_args_testee.with_bound_args(inexistent_arg=1)
-
-
-def test_call_bound_program_with_wrong_args(cartesian_case, bound_args_testee):
-    program_with_bound_arg = bound_args_testee.with_bound_args(arg1=True)
-    out = cases.allocate(cartesian_case, bound_args_testee, "out")()
-
-    with pytest.raises(TypeError) as exc_info:
-        program_with_bound_arg.with_backend(cartesian_case.backend)(out, offset_provider={})
-
-    assert (
-        re.search(
-            "Function takes 2 positional arguments, but 1 were given.",
-            exc_info.value.__cause__.args[0],
-        )
-        is not None
-    )
-
-
-def test_call_bound_program_with_already_bound_arg(cartesian_case, bound_args_testee):
-    program_with_bound_arg = bound_args_testee.with_bound_args(arg2=True)
-    out = cases.allocate(cartesian_case, bound_args_testee, "out")()
-
-    with pytest.raises(TypeError) as exc_info:
-        program_with_bound_arg.with_backend(cartesian_case.backend)(
-            True, out, arg2=True, offset_provider={}
-        )
-
-    assert (
-        re.search(
-            "Parameter 'arg2' already set as a bound argument.", exc_info.value.__cause__.args[0]
-        )
-        is not None
-    )
-
-
 @pytest.mark.uses_origin
 def test_direct_fo_call_with_domain_arg(cartesian_case):
     @field_operator
@@ -327,6 +224,7 @@ def test_direct_fo_call_with_domain_arg(cartesian_case):
 
 
 @pytest.mark.uses_origin
+@pytest.mark.uses_tuple_returns
 def test_direct_fo_call_with_domain_arg_tuple_return(cartesian_case):
     @field_operator
     def testee(inp: IField) -> tuple[IField, IField]:
@@ -342,3 +240,48 @@ def test_direct_fo_call_with_domain_arg_tuple_return(cartesian_case):
     ref[1:-1] = inp.asnumpy()[1:-1]
 
     cases.verify(cartesian_case, testee, inp, out=out, domain={IDim: (1, size - 1)}, ref=(ref, ref))
+
+
+def test_missing_arg_field_operator(cartesian_case):
+    """Test that calling a field_operator without required args raises an error."""
+
+    @gtx.field_operator(backend=cartesian_case.backend)
+    def copy(a: IField) -> IField:
+        return a
+
+    a = cases.allocate(cartesian_case, copy, "a")()
+
+    with pytest.raises(errors.MissingArgumentError, match="'out'"):
+        _ = copy(a, offset_provider={})
+
+
+def test_missing_arg_scan_operator(cartesian_case):
+    """Test that calling a scan_operator without required args raises an error."""
+
+    @gtx.scan_operator(backend=cartesian_case.backend, axis=KDim, init=0.0, forward=True)
+    def sum(state: float, a: float) -> float:
+        return state + a
+
+    a = cases.allocate(cartesian_case, sum, "a")()
+
+    with pytest.raises(errors.MissingArgumentError, match="'out'"):
+        _ = sum(a, offset_provider={})
+
+
+def test_missing_arg_program(cartesian_case):
+    """Test that calling a program without required args raises an error."""
+
+    @gtx.field_operator
+    def copy(a: IField) -> IField:
+        return a
+
+    a = cases.allocate(cartesian_case, copy, "a")()
+    b = cases.allocate(cartesian_case, copy, cases.RETURN)()
+
+    with pytest.raises(errors.DSLError, match="Invalid call"):
+
+        @gtx.program(backend=cartesian_case.backend)
+        def copy_program(a: IField, b: IField) -> IField:
+            copy(a)
+
+        _ = copy_program(a, offset_provider={})
