+++ tests/next_tests/integration_tests/feature_tests/ffront_tests/test_compiled_program.py
@@ -215,6 +215,7 @@ def compile_testee_unstructured_no_jit(compile_testee_unstructured):
     return compile_testee_unstructured.with_compilation_options(enable_jit=False)
 
 
+@pytest.mark.uses_unstructured_shift
 def test_compile_unstructured(unstructured_case, compile_testee_unstructured):
     if unstructured_case.backend is None:
         pytest.skip("Embedded compiled program doesn't make sense.")
@@ -248,6 +249,7 @@ def skip_value_mesh_descriptor(exec_alloc_descriptor):
     return skip_value_mesh(exec_alloc_descriptor.allocator)
 
 
+@pytest.mark.uses_unstructured_shift
 def test_compile_unstructured_wrong_offset_provider(
     unstructured_case, compile_testee_unstructured_no_jit, skip_value_mesh_descriptor
 ):
@@ -271,6 +273,7 @@ def test_compile_unstructured_wrong_offset_provider(
         )
 
 
+@pytest.mark.uses_unstructured_shift
 def test_compile_unstructured_for_two_offset_providers(
     unstructured_case, compile_testee_unstructured_no_jit, skip_value_mesh_descriptor
 ):
@@ -359,6 +362,7 @@ def compile_variants_testee(cartesian_case, compile_variants_testee_not_compiled
     )
 
 
+@pytest.mark.uses_tuple_returns
 def test_compile_variants(cartesian_case, compile_variants_testee):
     # make sure the backend is never called
     object.__setattr__(compile_variants_testee, "backend", _raise_on_compile)
@@ -401,6 +405,7 @@ def test_compile_variants(cartesian_case, compile_variants_testee):
     assert np.allclose(out[1].ndarray, field_b.ndarray - 4.0)
 
 
+@pytest.mark.uses_tuple_returns
 def test_compile_variants_args_and_kwargs(cartesian_case, compile_variants_testee):
     # make sure the backend is never called
     object.__setattr__(compile_variants_testee, "backend", _raise_on_compile)
@@ -422,6 +427,7 @@ def test_compile_variants_args_and_kwargs(cartesian_case, compile_variants_teste
     assert np.allclose(out[1].ndarray, field_b.ndarray + 3.0)
 
 
+@pytest.mark.uses_tuple_returns
 def test_compile_variants_not_compiled(cartesian_case, compile_variants_testee):
     object.__setattr__(compile_variants_testee.compilation_options, "enable_jit", False)
 
@@ -441,6 +447,7 @@ def test_compile_variants_not_compiled(cartesian_case, compile_variants_testee):
         )
 
 
+@pytest.mark.uses_tuple_returns
 def test_compile_variants_not_compiled_but_jit_enabled_on_call(
     cartesian_case, compile_variants_testee
 ):
@@ -531,6 +538,7 @@ def test_compile_variants_config_default_disable_jit(cartesian_case):
             )
 
 
+@pytest.mark.uses_tuple_returns
 def test_compile_variants_not_compiled_then_reset_static_params(
     cartesian_case, compile_variants_testee
 ):
@@ -589,6 +597,7 @@ def test_compile_variants_not_compiled_then_reset_static_params(
     assert np.allclose(out[1].ndarray, field_b.ndarray + 5.0)
 
 
+@pytest.mark.uses_tuple_returns
 def test_compile_variants_not_compiled_then_set_new_static_params(
     cartesian_case, compile_variants_testee
 ):
@@ -648,6 +657,7 @@ def test_compile_variants_not_compiled_then_set_new_static_params(
         )
 
 
+@pytest.mark.uses_tuple_returns
 def test_compile_variants_jit(cartesian_case, compile_variants_testee):
     object.__setattr__(compile_variants_testee.compilation_options, "enable_jit", True)
 
@@ -683,6 +693,7 @@ def test_compile_variants_jit(cartesian_case, compile_variants_testee):
     assert np.allclose(out[1].ndarray, field_b.ndarray - 4.0)
 
 
+@pytest.mark.uses_tuple_returns
 def test_compile_variants_with_static_params_jit(
     cartesian_case, compile_variants_testee_not_compiled
 ):
@@ -724,6 +735,7 @@ def test_compile_variants_with_static_params_jit(
     assert np.allclose(out[1].ndarray, field_b.ndarray + 3.0)
 
 
+@pytest.mark.uses_tuple_returns
 def test_compile_variants_decorator_static_params_jit(
     compile_variants_field_operator, cartesian_case
 ):
@@ -831,6 +843,7 @@ def compile_variants_testee_tuple(cartesian_case):
     )
 
 
+@pytest.mark.uses_tuple_args
 def test_compile_variants_tuple(cartesian_case, compile_variants_testee_tuple):
     # make sure the backend is never called
     object.__setattr__(compile_variants_testee_tuple, "backend", _raise_on_compile)
@@ -896,6 +909,8 @@ def test_wait_for_compilation(cartesian_case, compile_testee, compile_testee_dom
         compile_testee_domain.compile(offset_provider=cartesian_case.offset_provider)
 
 
+@pytest.mark.uses_tuple_args
+@pytest.mark.uses_tuple_returns
 def test_compile_variants_decorator_static_domains(cartesian_case):
     if cartesian_case.backend is None:
         pytest.skip("Embedded compiled program doesn't make sense.")
