+++ tests/next_tests/integration_tests/feature_tests/ffront_tests/test_program.py
@@ -130,6 +130,7 @@ def test_calling_fo_from_fo_execution(cartesian_case):
     )
 
 
+@pytest.mark.uses_tuple_returns
 def test_tuple_program_return_constructed_inside(cartesian_case):
     @gtx.field_operator
     def pack_tuple(
@@ -157,6 +158,7 @@ def test_tuple_program_return_constructed_inside(cartesian_case):
 
 
 @pytest.mark.uses_program_with_sliced_out_arguments
+@pytest.mark.uses_tuple_returns
 def test_tuple_program_return_constructed_inside_with_slicing(cartesian_case):
     @gtx.field_operator
     def pack_tuple(
@@ -186,6 +188,7 @@ def test_tuple_program_return_constructed_inside_with_slicing(cartesian_case):
     assert out_a[0].as_scalar() == 0 and out_b[0].as_scalar() == 0
 
 
+@pytest.mark.uses_tuple_returns
 def test_tuple_program_return_constructed_inside_nested(cartesian_case):
     @gtx.field_operator
     def pack_tuple(
