+++ tests/next_tests/integration_tests/feature_tests/ffront_tests/test_reductions.py
@@ -6,7 +6,6 @@
 # Please, refer to the LICENSE file in the root directory.
 # SPDX-License-Identifier: BSD-3-Clause
 
-import functools
 from typing import TypeAlias
 
 import numpy as np
@@ -23,8 +22,6 @@ from next_tests.integration_tests.cases import (
     V2E,
     E2VDim,
     Edge,
-    IDim,
-    Ioff,
     JDim,
     Joff,
     KDim,
@@ -138,6 +135,7 @@ def test_neighbor_sum(unstructured_case_3d, fop):
 
 
 @pytest.mark.uses_unstructured_shift
+@pytest.mark.uses_cartesian_shift
 def test_reduction_execution_with_offset(unstructured_case_3d):
     EKField: TypeAlias = gtx.Field[[Edge, KDim], np.int32]
     VKField: TypeAlias = gtx.Field[[Vertex, KDim], np.int32]
@@ -305,124 +303,6 @@ def test_reduction_expression_with_where_and_scalar(unstructured_case):
             initial=0,
             where=v2e_table != common._DEFAULT_SKIP_VALUE,
         ),
-    )
-
-
-@pytest.mark.uses_tuple_returns
-def test_conditional_nested_tuple(cartesian_case):
-    @gtx.field_operator
-    def conditional_nested_tuple(
-        mask: cases.IBoolField, a: cases.IFloatField, b: cases.IFloatField
-    ) -> tuple[
-        tuple[cases.IFloatField, cases.IFloatField], tuple[cases.IFloatField, cases.IFloatField]
-    ]:
-        return where(mask, ((a, b), (b, a)), ((5.0, 7.0), (7.0, 5.0)))
-
-    size = cartesian_case.default_sizes[IDim]
-    mask = cartesian_case.as_field([IDim], np.random.choice(a=[False, True], size=size))
-    a = cases.allocate(cartesian_case, conditional_nested_tuple, "a")()
-    b = cases.allocate(cartesian_case, conditional_nested_tuple, "b")()
-
-    where_with_mask = functools.partial(np.where, mask.asnumpy())
-
-    cases.verify(
-        cartesian_case,
-        conditional_nested_tuple,
-        mask,
-        a,
-        b,
-        out=cases.allocate(cartesian_case, conditional_nested_tuple, cases.RETURN)(),
-        ref=(
-            (
-                where_with_mask(a.asnumpy(), np.full(size, 5.0)),
-                where_with_mask(b.asnumpy(), np.full(size, 7.0)),
-            ),
-            (
-                where_with_mask(b.asnumpy(), np.full(size, 7.0)),
-                where_with_mask(a.asnumpy(), np.full(size, 5.0)),
-            ),
-        ),
-    )
-
-
-def test_conditional(cartesian_case):
-    @gtx.field_operator
-    def conditional(
-        mask: cases.IBoolField, a: cases.IFloatField, b: cases.IFloatField
-    ) -> cases.IFloatField:
-        return where(mask, a, b)
-
-    size = cartesian_case.default_sizes[IDim]
-    mask = cartesian_case.as_field([IDim], np.random.choice(a=[False, True], size=(size)))
-    a = cases.allocate(cartesian_case, conditional, "a")()
-    b = cases.allocate(cartesian_case, conditional, "b")()
-    out = cases.allocate(cartesian_case, conditional, cases.RETURN)()
-
-    cases.verify(
-        cartesian_case,
-        conditional,
-        mask,
-        a,
-        b,
-        out=out,
-        ref=np.where(mask.asnumpy(), a.asnumpy(), b.asnumpy()),
-    )
-
-
-def test_conditional_promotion(cartesian_case):
-    @gtx.field_operator
-    def conditional_promotion(mask: cases.IBoolField, a: cases.IFloatField) -> cases.IFloatField:
-        return where(mask, a, 10.0)
-
-    size = cartesian_case.default_sizes[IDim]
-    mask = cartesian_case.as_field([IDim], np.random.choice(a=[False, True], size=(size)))
-    a = cases.allocate(cartesian_case, conditional_promotion, "a")()
-    out = cases.allocate(cartesian_case, conditional_promotion, cases.RETURN)()
-    ref = np.where(mask.asnumpy(), a.asnumpy(), 10.0)
-
-    cases.verify(cartesian_case, conditional_promotion, mask, a, out=out, ref=ref)
-
-
-def test_conditional_compareop(cartesian_case):
-    @gtx.field_operator
-    def conditional_promotion(a: cases.IFloatField) -> cases.IFloatField:
-        return where(a != a, a, 10.0)
-
-    cases.verify_with_default_data(
-        cartesian_case, conditional_promotion, ref=lambda a: np.where(a != a, a, 10.0)
-    )
-
-
-@pytest.mark.uses_cartesian_shift
-def test_conditional_shifted(cartesian_case):
-    @gtx.field_operator
-    def conditional_shifted(
-        mask: cases.IBoolField, a: cases.IFloatField, b: cases.IFloatField
-    ) -> gtx.Field[[IDim], float64]:
-        tmp = where(mask, a, b)
-        return tmp(Ioff[1])
-
-    @gtx.program
-    def conditional_program(
-        mask: cases.IBoolField, a: cases.IFloatField, b: cases.IFloatField, out: cases.IFloatField
-    ):
-        conditional_shifted(mask, a, b, out=out)
-
-    size = cartesian_case.default_sizes[IDim] + 1
-    mask = cartesian_case.as_field([IDim], np.random.choice(a=[False, True], size=(size)))
-    a = cases.allocate(cartesian_case, conditional_program, "a").extend({IDim: (0, 1)})()
-    b = cases.allocate(cartesian_case, conditional_program, "b").extend({IDim: (0, 1)})()
-    out = cases.allocate(cartesian_case, conditional_shifted, cases.RETURN)()
-
-    cases.verify(
-        cartesian_case,
-        conditional_program,
-        mask,
-        a,
-        b,
-        out,
-        inout=out,
-        ref=np.where(mask.asnumpy(), a.asnumpy(), b.asnumpy())[1:],
     )
 
 
@@ -646,6 +526,7 @@ def test_ternary_builtin_neighbor_sum(unstructured_case):
     )
 
 
+@pytest.mark.uses_unstructured_shift
 def test_local_index_premapped_field(request, unstructured_case):
     if request.node.get_closest_marker(pytest.mark.uses_mesh_with_skip_values.name):
         pytest.skip("This test only works with non-skip value meshes.")
