+++ tests/next_tests/integration_tests/feature_tests/iterator_tests/test_program.py
@@ -88,6 +88,7 @@ def index_program_shift(out, size):
 
 
 @pytest.mark.uses_index_fields
+@pytest.mark.uses_cartesian_shift
 def test_index_builtin_shift(program_processor):
     program_processor, validate = program_processor
